BreakPoint

+++ Source/PuzzlePlatforms/PlayersComponent/MyCharacterStatComponent.h
@@ -79,7 +79,7 @@ public:
 	UPROPERTY(EditAnywhere, Transient)
 		float CurrentMP;
 	int32 Level;
-	int AttackDamage;
+	float AttackDamage;
 	FText Name;
 	int CurrentExp = 0;
 	FMyCharacterrData* CurrentStatData = nullptr;
